refactor : common move to postWrite components
