+++ frontend/src/components/PostWrite/common/Content.js
