+++ frontend/src/components/PostWrite/common/FlexBox.js
