+++ frontend/src/components/PostWrite/common/SelectButton.js
@@ -1,16 +1,21 @@
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import { device } from "styles/media";
 import Button from "components/common/Button";
 
 const SelectButton = styled(Button)`
-  border: 1px solid #aaaaaa;
   background-color: ${({ theme }) => theme.colors.oppositeDefalut};
   font-size: 0.7rem;
   height: 30px;
   margin-left: 3px;
   width: 50px;
-  border: 1px solid ${(props) => props?.color};
-  color: ${(props) => props?.color};
+  color: ${({ theme }) => theme.colors.nonSelected};
+  border: 1px solid ${({ theme }) => theme.colors.nonSelected};
+  ${(props) =>
+    props.isSelected &&
+    css`
+      color: ${({ theme }) => theme.colors.selected};
+      border: 1px solid ${({ theme }) => theme.colors.selected};
+    `}
 
   @media ${device.tablet} {
     font-size: 1.3em;
