+++ frontend/src/components/PostWrite/common/StyledInput.js
