+++ frontend/src/components/PostWrite/common/StyledInputMoney.js
