+++ frontend/src/components/PostWrite/common/index.js
@@ -4,13 +4,14 @@ import ContentBody from "./ContentBody";
 import FlexBox from "./FlexBox";
 import InlineSubTitle from "./InlineSubTitle";
 import Manwon from "./Manwon";
-import PrevNextBtn from "./PrevNext";
+import PrevNextBtn from "./PrevNextBtn";
 import SelectButton from "./SelectButton";
 import SpaceBetween from "./SpaceBetween";
 import StyledInput from "./StyledInput";
 import StyledInputMoney from "./StyledInputMoney";
 import StyledInputDate from "./StyledInputDate";
 import SubTitle from "./SubTitle";
+import InputMoney from "./InputMoney";
 
 export {
   Title,
@@ -23,6 +24,7 @@ export {
   PrevNextBtn,
   SelectButton,
   SpaceBetween,
+  InputMoney,
   StyledInput,
   StyledInputDate,
   StyledInputMoney,
