+++ frontend/src/pages/PostWrite/PostWrite.js
@@ -9,7 +9,7 @@ import PhaseFour from "components/PostWrite/PhaseFour";
 import PhaseFive from "components/PostWrite/PhaseFive";
 import PhaseSix from "components/PostWrite/PhaseSix";
 import WriteProgressBar from "components/common/WriteProgressBar";
-import PrevNext from "components/PostWrite/PrevNext";
+import PrevNext from "components/PostWrite/common/PrevNext";
 
 const PostWritePage = () => {
   return (
